requirements.txt
fastapi
uvicorn
python-multipart
google-auth
google-auth-oauthlib
google-api-python-client
google-generativeai
pandas
python-dotenv
aiofiles
google-generativeai
